+++ .env.example
@@ -1,0 +1,3 @@
+MICROCMS_API_DOMAIN=
+MICROCMS_API_KEY=
+NEXT_PUBLIC_CODE_API_URL=/api/mock/codes
