+++ types/types.d.ts
@@ -1,8 +1,17 @@
+type Theme = 'default' | 'dracula' | 'sunset' | 'dim'
+
 type Line = {
     id: string
     line: string
 }
 
-type Theme = 'default' | 'dracula' | 'sunset' | 'dim'
-
 type Code = Line[]
+
+type MicroCMSContents = {
+    id: string
+    createdAt: string
+    updatedAt: string
+    publishedAt: string
+    revisedAt: string
+    text: string
+}
